change the datatype of the dc values to 16 bits
git-svn-id: svn://svn.mplayerhq.hu/soc/dirac@844 66570020-b32e-0410-962d-e44e8f9afd43

+++ libavcodec/dirac.c
@@ -216,7 +216,7 @@ struct dirac_blockmotion {
     uint8_t use_ref[2];
     uint8_t use_global;
     vect_t vect[2];
-    int8_t dc[3];
+    int16_t dc[3];
 };
 
 /* XXX */
